Add parse date and time
